normal matches can be generated.
Already generated schedules do not get send to the api anymore. Faster opening of widgets.
winner can be set and is visible in the timetable
working on finals right now

found issues:
function create pools needs to be placed to another end. Because at the moment if the user would swap teams, they would not be send to the api. So the teams do not get swapped, even if the user did it